requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
yt-dlp==2025.4.30
ffmpeg-python==0.2.0
boto3==1.36.14
celery==5.3.1
pydub==0.25.1
assemblyai==0.37.0